back-end/requirements.txt
fastapi
sqlalchemy
pydantic
Uvicorn
PyPDF2
groq
spacy
langchain
flashtext
langchain_groq
python-dotenv
requests
pywsd
nltk==3.6.7
langchain-community
langchain-huggingface
transformers
sentence-transformers
torch
pytorch_pretrained_bert
python-multipart
git+https://github.com/boudinfl/pke.git
